+++ pkg/nodeagent/apis/config/v1alpha1/validation/validation_suite_test.go
@@ -13,5 +13,5 @@ import (
 
 func TestValidation(t *testing.T) {
 	RegisterFailHandler(Fail)
-	RunSpecs(t, "NodeAgent APIs Config Validation Suite")
+	RunSpecs(t, "NodeAgent APIs Config V1alpha1 Validation Suite")
 }
