+++ pkg/nodeagent/apis/config/v1alpha1/validation/validation_test.go
@@ -14,8 +14,8 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/validation/field"
 
-	. "github.com/gardener/gardener/pkg/nodeagent/apis/config"
-	. "github.com/gardener/gardener/pkg/nodeagent/apis/config/validation"
+	. "github.com/gardener/gardener/pkg/nodeagent/apis/config/v1alpha1"
+	. "github.com/gardener/gardener/pkg/nodeagent/apis/config/v1alpha1/validation"
 )
 
 var _ = Describe("#ValidateNodeAgentConfiguration", func() {
